ERT-1059 Moved WorkflowDialog
